requirements.txt
elasticsearch
absl-py==0.11.0
astor==0.8.1
bert-serving-client==1.10.0
bert-serving-server==1.9.8
cached-property==1.5.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
gast==0.2.2
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.36.0
gunicorn==20.0.4
h5py==3.1.0
importlib-metadata==3.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.19.5
opt-einsum==3.3.0
protobuf==3.15.3
PyJWT==1.4.2
pyzmq==22.0.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
